+++ src/main/java/de/maxhenkel/audioplayer/ComponentUtils.java
@@ -10,12 +10,13 @@ import net.minecraft.world.item.EitherHolder;
 import net.minecraft.world.item.Instrument;
 import net.minecraft.world.item.JukeboxPlayable;
 import net.minecraft.world.item.JukeboxSong;
+import net.minecraft.world.item.component.InstrumentComponent;
 
 public class ComponentUtils {
 
-    public static final Holder<Instrument> EMPTY_INSTRUMENT = Holder.direct(new Instrument(Holder.direct(SoundEvents.EMPTY), 140, 256F, Component.empty()));
+    public static final InstrumentComponent EMPTY_INSTRUMENT = new InstrumentComponent(Holder.direct(new Instrument(Holder.direct(SoundEvents.EMPTY), 140, 256F, Component.empty())));
 
     public static final ResourceKey<JukeboxSong> CUSTOM_JUKEBOX_SONG_KEY = ResourceKey.create(Registries.JUKEBOX_SONG, ResourceLocation.fromNamespaceAndPath(AudioPlayer.MODID, "custom"));
-    public static final JukeboxPlayable CUSTOM_JUKEBOX_PLAYABLE = new JukeboxPlayable(new EitherHolder<>(CUSTOM_JUKEBOX_SONG_KEY), false);
+    public static final JukeboxPlayable CUSTOM_JUKEBOX_PLAYABLE = new JukeboxPlayable(new EitherHolder<>(CUSTOM_JUKEBOX_SONG_KEY));
 
 }