+++ src/main/java/de/maxhenkel/audioplayer/mixin/SkullBlockEntityMixin.java
@@ -6,6 +6,7 @@ import de.maxhenkel.audioplayer.interfaces.ChannelHolder;
 import de.maxhenkel.audioplayer.interfaces.CustomSoundHolder;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.UUIDUtil;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -56,7 +57,7 @@ public class SkullBlockEntityMixin extends BlockEntity implements CustomSoundHol
     @Inject(method = "saveAdditional", at = @At("RETURN"))
     private void saveAdditional(CompoundTag tag, HolderLookup.Provider provider, CallbackInfo ci) {
         if (channelID != null) {
-            tag.putUUID("ChannelID", channelID);
+            tag.store("ChannelID", UUIDUtil.CODEC, channelID);
         }
         if (customSound != null) {
             customSound.saveToNbt(tag);
@@ -65,11 +66,7 @@ public class SkullBlockEntityMixin extends BlockEntity implements CustomSoundHol
 
     @Inject(method = "loadAdditional", at = @At("RETURN"))
     private void load(CompoundTag tag, HolderLookup.Provider provider, CallbackInfo ci) {
-        if (tag.contains("ChannelID")) {
-            channelID = tag.getUUID("ChannelID");
-        } else {
-            channelID = null;
-        }
+        channelID = tag.read("ChannelID", UUIDUtil.CODEC).orElse(null);
         customSound = CustomSound.of(tag);
     }