+++ changelog.md
@@ -1,1 +1,1 @@
-- Update volume command to make it perceptually linear
+- Fixed permissions not working from the console when using a permission manager