+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed permissions not working from the console when using a permission manager
+- Updated to 1.21.5